Systematically use relative_url filter for links
This allows us to deploy with a base URL which is helpful for preview deployments. It's not ideal but it's a pragmatic way to sort out our review deployment setup.

This also upgrades Jekyll to the latest 3.x for relative_url support.
Note if you don't install the latest it just silently fails to work.